Add transition guards to prevent premature rebounds
